// src/enums/sharedEvent.enum.ts
export enum Category {
   RACE = 'RACE',
   WORKOUT = 'WORKOUT'
}

export enum Type {
   Ride = 'Ride',
   Run = 'Run',
   Swim = 'Swim',
   WeightTraining = 'WeightTraining',
   Hike = 'Hike',
   Walk = 'Walk',
   AlpineSki = 'AlpineSki',
   BackcountrySki = 'BackcountrySki',
   Badminton = 'Badminton',
   Canoeing = 'Canoeing',
   Crossfit = 'Crossfit',
   EBikeRide = 'EBikeRide',
   EMountainBikeRide = 'EMountainBikeRide',
   Elliptical = 'Elliptical',
   Golf = 'Golf',
   GravelRide = 'GravelRide',
   Handcycle = 'Handcycle',
   HighIntensityIntervalTraining = 'HighIntensityIntervalTraining',
   Hockey = 'Hockey',
   IceSkate = 'IceSkate',
   InlineSkate = 'InlineSkate',
   Kayaking = 'Kayaking',
   Kitesurf = 'Kitesurf',
   MountainBikeRide = 'MountainBikeRide',
   NordicSki = 'NordicSki',
   OpenWaterSwim = 'OpenWaterSwim',
   Padel = 'Padel',
   Pilates = 'Pilates',
   Pickleball = 'Pickleball',
   Racquetball = 'Racquetball',
   Rugby = 'Rugby',
   RockClimbing = 'RockClimbing',
   RollerSki = 'RollerSki',
   Rowing = 'Rowing',
   Sail = 'Sail',
   Skateboard = 'Skateboard',
   Snowboard = 'Snowboard',
   Snowshoe = 'Snowshoe',
   Soccer = 'Soccer',
   Squash = 'Squash',
   StairStepper = 'StairStepper',
   StandUpPaddling = 'StandUpPaddling',
   Surfing = 'Surfing',
   TableTennis = 'TableTennis',
   Tennis = 'Tennis',
   TrailRun = 'TrailRun',
   Transition = 'Transition',
   Velomobile = 'Velomobile',
   VirtualRide = 'VirtualRide',
   VirtualRow = 'VirtualRow',
   VirtualRun = 'VirtualRun',
   WaterSport = 'WaterSport',
   Wheelchair = 'Wheelchair',
   Windsurf = 'Windsurf',
   Workout = 'Workout',
   Yoga = 'Yoga',
   Other = 'Other'
}

export enum Visibility {
   PUBLIC = 'PUBLIC',
   GROUP = 'GROUP'
}